symlink -- BUG
I think the kernel has a bug.
which the symlinktest revelead.

two processes that are forked from the same father.
the first one unlikns some dir entry while the dir entry still have file
ref > 1.

Later on when the process decreases the refs. [dont know if this is
related or not]
and the second process tries to open the dir entry, it finds it for some
reason.