Improved buffer cache
added a hash map with a hardcoded 31 slots :).
and a free buf list, each process tries to lock some specific hash map
entry, and if finds the buf it returns as before.

if enty is not found, we need to use one of the free available buff.

At the init phase we add all buffers to the vacant list, and inorder to
get on of these buffer we do cas to pick one buff, and cas to return it
once we are done.